requirements.txt
fastapi
uvicorn
joblib
scikit-learn
numpy
pandas
pydantic
cloudpickle
python-dotenv
google-generativeai
